+++ Source/Modules/XWrite/XWrite.psm1
@@ -2,7 +2,7 @@
 
 .DESCRIPTION PowerShell module for XWrite
 
-.VERSION 1.0
+.VERSION 1.1
 
 #>