+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+**1.1**
+
+- New segment in output `-Source` or `%source%` which can be 
+  - `Script`
+  - `Function`
+  - a module name
+- Fix `Enable-XWrite` parameter set names
+- Parameter `-Caller` is removed. Caller will be always outputed.
+
 **1.0**
 
 XWrite PowerShell module introduced